Interaction Media Project - Camera Software Tool
Create a software tool that records live interaction inputs as time series data (at least 20 samples of
recording, but more likely hundreds) before transforming the data into a static or dynamic graphic,
visual, auditory or haptic form for one or more media of your choice. You should provide feedback in
some way when the interaction is being recorded to the time series, as well as presenting a final form
that represents the accumulated interaction inputs.